Rename some aliases, update README and flake.lock
KeePassXC-Browser is the official name of the project and it was kinda
weird seeing Firefox become dependent on what otherwise looked like
regular KeePassXC, so rename to `keepassxc-browser` for more
distinction.

Rename Catppuccin file icons and containers to their official slugs.

Simplify the passthrough handling in the overlay since I forgot to do
that yesterday after I simplified it in the updater.

Update README to give an example on how to configure policies.